Refactor OsmApi class into different files
