+++ tests/osmapi_test.py
@@ -2,12 +2,7 @@ from __future__ import unicode_literals
 from osmapi import OsmApi
 import mock
 import os
-import sys
-
-if sys.version_info < (2, 7):
-    import unittest2 as unittest
-else:
-    import unittest
+import unittest
 
 __location__ = os.path.realpath(
     os.path.join(
@@ -28,10 +23,6 @@ class TestOsmApi(unittest.TestCase):
         print(self.api)
 
     def _session_mock(self, auth=False, filenames=None, status=200):
-        if auth:
-            self.api._username = 'testuser'
-            self.api._password = 'testpassword'
-
         response_mock = mock.Mock()
         response_mock.status_code = status
         return_values = self._return_values(filenames)
@@ -41,13 +32,24 @@ class TestOsmApi(unittest.TestCase):
         if return_values:
             response_mock.content = return_values[0]
 
-        session_mock = mock.Mock()
-        session_mock.request = mock.Mock(return_value=response_mock)
+        self.session_mock = mock.Mock()
+        self.session_mock.request = mock.Mock(return_value=response_mock)
 
-        self.api._get_http_session = mock.Mock(return_value=session_mock)
-        self.api._session = session_mock
+        if auth:
+            self.api = OsmApi(
+                api=self.api_base,
+                username='testuser',
+                password='testpassword',
+                session=self.session_mock
+            )
+        else:
+            self.api = OsmApi(
+                api=self.api_base,
+                session=self.session_mock
+            )
 
-        self.api._sleep = mock.Mock()
+        self.api._get_http_session = mock.Mock(return_value=self.session_mock)
+        self.api._session._sleep = mock.Mock()
 
     def _return_values(self, filenames):
         if filenames is None:
